updated handling of servo status errors
the way error status was handled in the driver was not quite so nice,
since it was not possible to differentiate between communication errors
and status errors of the servo (e.g. overload, overtemperature a.s.o)

I've added an ErrorStatus struct, which is set based on the return packet
from the dynamixel. Reading commands will still return true, even if this
status is not ok. Writing commands will still return 0 if the the error
servo value is false.